backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
pydantic-settings>=2.6.0
redis>=5.2.0
asyncpg>=0.30.0
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-cov>=6.0.0
httpx>=0.28.0
